+++ R/check.r
@@ -40,11 +40,12 @@
 #'   \code{\link{as.package}} for more information
 #' @param document if \code{TRUE} (the default), will update and check
 #'   documentation before running formal check.
-#' @param cleanup if \code{TRUE} the check directory is removed if the check
-#'   is successful - this allows you to inspect the results to figure out what
-#'   went wrong. If \code{FALSE} the check directory is never removed.
+#' @param cleanup Deprecated.
 #' @param cran if \code{TRUE} (the default), check using the same settings as
 #'   CRAN uses.
+#' @param run_dont_test Sets \code{--run-donttest} so that tests surrounded in
+#'   \code{\\dontest\{\}} are also tested. This is important for CRAN
+#'   submission.
 #' @param check_version Sets \code{_R_CHECK_CRAN_INCOMING_} env var.
 #'   If \code{TRUE}, performns a number of checked related
 #'   to version numbers of packages on CRAN.
@@ -60,34 +61,46 @@
 #'   CRAN.
 #' @export
 check <- function(pkg = ".", document = TRUE, cleanup = TRUE, cran = TRUE,
-                  check_version = FALSE, force_suggests = FALSE, args = NULL,
-                  build_args = NULL, quiet = FALSE, check_dir = tempdir(),
-                  ...) {
-
+                  check_version = FALSE, force_suggests = FALSE,
+                  run_dont_test = FALSE, args = NULL, build_args = NULL,
+                  quiet = FALSE, check_dir = tempdir(), ...) {
   pkg <- as.package(pkg)
+  if (!missing(cleanup)) {
+    warning("`cleanup` is deprecated", call. = FALSE)
+  }
 
   if (document) {
     document(pkg)
   }
 
-  show_env_vars(compiler_flags(FALSE))
-  withr::with_envvar(compiler_flags(FALSE), {
-
+  if (!quiet) {
+    show_env_vars(compiler_flags(FALSE))
     rule("Building ", pkg$package)
+  }
+  withr::with_envvar(compiler_flags(FALSE), action = "prefix", {
     built_path <- build(pkg, tempdir(), quiet = quiet, args = build_args, ...)
     on.exit(unlink(built_path), add = TRUE)
+  })
 
-    r_cmd_check_path <- check_r_cmd(pkg$package, built_path, cran, check_version,
-      force_suggests, args, quiet = quiet, check_dir = check_dir)
+  check_path <- check_r_cmd(
+    pkg$package,
+    built_path = built_path,
+    cran = cran,
+    check_version = check_version,
+    force_suggests = force_suggests,
+    run_dont_test = run_dont_test,
+    args = args,
+    quiet = quiet,
+    check_dir = check_dir
+  )
 
-    if (cleanup) {
-      unlink(r_cmd_check_path, recursive = TRUE)
-    } else {
-      if (!quiet) message("R CMD check results in ", r_cmd_check_path)
-    }
+  if (!quiet) {
+    rule("Check summary")
+    results <- parse_check_results(file.path(check_path, "00check.log"))
+    print(results)
+  }
 
-    invisible(TRUE)
-  }, "prefix")
+  invisible(check_path)
 }
 
 
@@ -96,7 +109,8 @@ check <- function(pkg = ".", document = TRUE, cleanup = TRUE, cran = TRUE,
 # @param check_dir The directory to unpack the .tar.gz file to
 check_r_cmd <- function(name, built_path = NULL, cran = TRUE,
                         check_version = FALSE, force_suggests = FALSE,
-                        args = NULL, check_dir = tempdir(), quiet = FALSE, ...) {
+                        run_dont_test = FALSE, args = NULL,
+                        check_dir = tempdir(), quiet = FALSE, ...) {
 
   pkgname <- gsub("_.*?$", "", basename(built_path))
 
@@ -107,15 +121,18 @@ check_r_cmd <- function(name, built_path = NULL, cran = TRUE,
     opts <- c(opts, "--no-build-vignettes", "--no-manual")
   }
   if (cran) {
-    opts <- c("--as-cran", "--run-donttest", opts)
+    opts <- c("--as-cran", opts)
+  }
+  if (run_dont_test) {
+    opts <- c("--run-donttest", opts)
   }
 
   env_vars <- check_env_vars(cran, check_version, force_suggests)
-  if (!quiet)
+  if (!quiet) {
     show_env_vars(env_vars)
-
-  if (!quiet)
     rule("Checking ", name)
+  }
+
   opts <- paste(paste(opts, collapse = " "), paste(args, collapse = " "))
   R(paste("CMD check ", shQuote(built_path), " ", opts, sep = ""), check_dir,
     env_vars, quiet = quiet, ...)